Filled out remaining pages and cleaned out code
