+++ sea_project/assets/custom.css
@@ -12069,3 +12069,111 @@ textarea.form-control-lg {
 .no-margin-p p {
   margin-bottom: 0 !important;
 }
+
+
+
+
+
+/* Collage container setup */
+
+.intro-section {
+    position: relative;
+    height: 600px;
+    margin-bottom: 2rem;
+}
+
+.collage-container {
+    position: relative;
+    height: 100%;
+    width: 100%;
+}
+
+/* Banner centered in the collage */
+.banner-text {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 2rem;
+    font-weight: bold;
+    color: white;
+    background-color: rgba(0, 0, 0, 0.6);
+    padding: 1rem 2rem;
+    border-radius: 10px;
+    z-index: 10;
+    white-space: nowrap;
+}
+.collage-slot {
+    position: absolute;
+    width: 300px;
+    max-width: 25%;
+    text-align: center;
+}
+
+.img-label {
+    margin-bottom: 0.5rem;
+    font-weight: 600;
+    font-size: 1rem;
+    background-color: --bs-dark-bg;
+    padding: 0.2rem 0.5rem;
+    border-radius: 5px;
+    display: inline-block;
+}
+
+.collage-img {
+    width: 100%;
+    max-width: 400px;
+    border-radius: 10px;
+    box-shadow: 2px 2px 10px rgba(0,0,0,0.3);
+    transition: opacity 1s ease-in-out;
+    background-color: transparent;
+}
+
+/* Image positions */
+.img1 { top: 0%; left: 12%; z-index: 3; }
+.img2 { top: 15%; left: 30%; z-index: 2; }
+.img3 { top: 5%; left: 60%; z-index: 4; }
+.img4 { top: 50%; left: 0%; z-index: 1; }
+.img5 { top: 55%; left: 45%; z-index: 2; }
+.img6 { top: 35%; left: 75%; z-index: 3; }
+
+
+
+/* Responsive layout for smaller screens */
+@media (max-width: 768px) {
+    .collage-container {
+        height: auto;
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: center;
+    }
+
+    .collage-img {
+        position: static;
+        width: 80%;
+        height: auto;
+        margin: 1rem;
+    }
+
+    .banner-text {
+        position: static;
+        transform: none;
+        margin: 1rem auto;
+    }
+}
+
+
+/* Project styling */
+.project-img {
+    width: 100%;
+    height: 100%;
+    max-height: 400px;
+    max-width: 400px;
+    object-fit: cover;
+    border-radius: 12px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
+    background-color: transparent;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
